feat(recipient-picker): one address combobox for transfer and reward payout
