+++ src/renderer/widgets/RecipientPicker/index.ts
@@ -1,0 +1,3 @@
+export { RecipientCombobox } from './ui/RecipientCombobox';
+export { type RecipientGroup, type RecipientOption, useRecipientOptions } from './lib/useRecipientOptions';
+export { filterRecipientAccounts } from './lib/recipient-accounts';
